Switch to YACL (#35)
* add YACL

* Revert "Remove client-side code (#28)"

This reverts commit 554f94fea94676dc90a17c4cacabb82ed240c1e5.

* add modmenu implementation

* remove useless client code

* define config in yacl config api
* add empty config screen
* add time category
* add gametime category
* fix wrong value being set
* add debug category

* remove comments from old config
* migrate time config
* migrate gametime config
* migrate debug config

* remove old config class
* remove leftover comment

* rename config class

* remove unused import
* empty line cus idk

* remove midnightlib import
* remove dependency for midnightlib
* remove unused translation strings

* use translation strings instead of literal text values

* make internal fields final where they could be

* declare dependency on YACL

+++ src/main/java/xyz/verarr/synchrono/IRLTimeManager.java
@@ -9,6 +9,7 @@ import java.time.*;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
 import java.util.function.Function;
+
 import xyz.verarr.synchrono.external_apis.SunriseSunsetAPI;
 import xyz.verarr.synchrono.external_apis.SunriseSunsetAPI.SunriseSunsetData;
 
@@ -105,7 +106,7 @@ public class IRLTimeManager extends PersistentState {
         if (!SynchronoConfig.invert) daytime_ticks = hardDaytimeTicksAt(dateTime);
         else daytime_ticks = hardNighttimeTicksAt(dateTime) + TICKS_PER_HALF_DAY;
 
-        daytime_ticks = Math.round(daytime_ticks / SynchronoConfig.scalar);
+        daytime_ticks = (int) Math.round(daytime_ticks / SynchronoConfig.scalar);
 
         return daytime_ticks;
     }
@@ -116,7 +117,7 @@ public class IRLTimeManager extends PersistentState {
         if (!SynchronoConfig.invert) nighttime_ticks = hardNighttimeTicksAt(dateTime);
         else nighttime_ticks = hardDaytimeTicksAt(dateTime) + TICKS_PER_HALF_DAY;
 
-        nighttime_ticks = Math.round(nighttime_ticks / SynchronoConfig.scalar);
+        nighttime_ticks = (int) Math.round(nighttime_ticks / SynchronoConfig.scalar);
 
         return nighttime_ticks;
     }
